# packages/shared/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nuru Place Pathway API
  version: "1.0.0"
  description: >
    Versioned REST surface over the §2 schema. Every endpoint is stateless,
    authenticated at the gateway, and idempotent wherever it creates state from a
    flaky client (spec §3). This document is the versioned contract validated in
    CI (§1.11, §3.7). The endpoint catalog below is the full v1 surface from §3.3;
    request/response schemas are filled in incrementally as features land — only
    the QuizSubmission/QuizResult slice from §3.7 is fleshed out so far.
servers:
  - url: https://api.nuruplace.org/v1
    description: Production (major version in path; additive-only within a major)
...
security:
  - bearerAuth: []
paths:
...
  /auth/oauth/{provider}:
    post:
      tags: [auth]
      summary: Exchange an IdP authorization code for our tokens; provisions on first login.
      security: []
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [kingschat, google, apple] } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OAuthExchange" }
      responses:
        "200":
          description: Token pair issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/UpstreamUnavailable" }
  /auth/token/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token; reuse detection revokes the family.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New access + refresh pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current refresh-token family.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204": { description: Logged out }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/dev-login:
    post:
      tags: [auth]
      summary: DEV ONLY. Mint a session for a seeded user by email/user_id. 404 in production (never mounted).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                user_id: { type: string, format: uuid }
      responses:
        "200":
          description: Token pair issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/mfa/enroll:
    post:
      tags: [auth]
      summary: Begin TOTP enrollment; returns an otpauth URI for an authenticator app (§5.3).
      responses:
        "201":
          description: Enrollment started (activates on first verify)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaEnroll" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/mfa/verify:
    post:
      tags: [auth]
      summary: Verify a TOTP code; returns an MFA-elevated access token for step-up (§5.3).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MfaVerify" }
      responses:
        "200":
          description: MFA-elevated access token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaElevation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me:
    get:
      tags: [auth]
      summary: Current user profile + enrollment summary + active level.
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [auth]
      summary: Update mutable profile fields with row_version check.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /me/onboarding:
    post:
      tags: [auth]
      summary: Submit baseline intake + literacy quiz; instantiates enrollment at L1·M1.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OnboardingIntake" }
      responses:
        "201": { description: Enrolled }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /me/devices:
    post:
      tags: [auth]
      summary: Register a device + push token; returns a device_id for sync cursors.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceRegister" }
      responses:
        "201": { description: Device registered }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # --- Curriculum & content (§3.3) ---
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Conflict:
      description: Optimistic-concurrency clash; current value returned.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: Well-formed but semantically invalid (e.g. unknown cell group).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UpstreamUnavailable:
      description: A required dependency (e.g. the IdP) is unreachable or not configured.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }

    # --- Auth & identity (§3.3, §5.3) ---
    OAuthExchange:
      type: object
      required: [code]
      description: Server-side authorization-code exchange inputs (§5.3).
      properties:
        code: { type: string, minLength: 1 }
        redirect_uri: { type: string, format: uri }
        code_verifier: { type: string, minLength: 1, description: PKCE verifier }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string, description: Short-lived (15-min) access JWT }
        refresh_token: { type: string, description: Rotating refresh token (returned once) }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: Access-token lifetime in seconds }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 1 }
    MeProfile:
      type: object
      required: [profile]
      properties:
        profile:
          type: object
          required: [user_id, full_name, role, congregation_id, row_version]
          properties:
            user_id: { type: string, format: uuid }
            email: { type: [string, "null"] }
            full_name: { type: string }
            phone_number: { type: [string, "null"] }
            date_of_birth: { type: [string, "null"], format: date }
            year_of_salvation: { type: [integer, "null"] }
            is_baptized: { type: boolean }
            cell_group_id: { type: [string, "null"], format: uuid }
            congregation_id: { type: [string, "null"], format: uuid }
            role: { type: string, enum: [Student, Instructor, Admin, SuperAdmin] }
            timezone: { type: string }
            locale: { type: string }
            is_minor: { type: boolean }
            row_version: { type: integer }
        enrollment:
          type: [object, "null"]
          properties:
            enrollment_id: { type: string, format: uuid }
            current_level: { type: integer }
            state: { type: string, enum: [active, paused, completed, withdrawn] }
            started_at: { type: string, format: date-time }
    ProfileUpdate:
      type: object
      required: [row_version]
      description: Mutable fields only; role/congregation are not self-writable (§5.4, §5.8).
      properties:
        phone_number: { type: string, minLength: 3, maxLength: 32 }
        cell_group_id: { type: [string, "null"], format: uuid }
        timezone: { type: string, maxLength: 64 }
        locale: { type: string, maxLength: 12 }
        row_version: { type: integer, description: Expected version for optimistic concurrency }
    OnboardingIntake:
      type: object
      required: [date_of_birth, phone_number, cell_group_id]
      properties:
        date_of_birth: { type: string, format: date }
        phone_number: { type: string, minLength: 3, maxLength: 32 }
        cell_group_id: { type: string, format: uuid }
        year_of_salvation: { type: integer, minimum: 1900, maximum: 2100 }
        is_baptized: { type: boolean, default: false }
        timezone: { type: string, maxLength: 64 }
    DeviceRegister:
      type: object
      required: [platform]
      properties:
        platform: { type: string, enum: [ios, android] }
        app_version: { type: string, maxLength: 20 }
        push_token: { type: string, maxLength: 512 }
    MfaEnroll:
      type: object
      required: [otpauth_uri, secret]
      properties:
        otpauth_uri: { type: string, description: otpauth:// URI for QR provisioning }
        secret: { type: string, description: base32 TOTP secret (shown once for manual entry) }
    MfaVerify:
      type: object
      required: [code]
      properties:
        code: { type: string, pattern: "^[0-9]{6,10}$" }
    MfaElevation:
      type: object
      required: [access_token, token_type, expires_in, mfa_enabled]
      properties:
        access_token: { type: string, description: Access token carrying the verified mfa claim }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }
        mfa_enabled: { type: boolean }
...